Updated documentation and changed count from int to long

+++ Net.Http.OData/ODataResponseContent.cs
@@ -13,12 +13,15 @@
 namespace Net.Http.OData
 {
     /// <summary>
-    /// A class which is used to return OData content.
+    /// A class which is used to return the OData content to an OData request.
     /// </summary>
+    /// <remarks>
+    /// <![CDATA[http://docs.oasis-open.org/odata/odata-json-format/v4.0/errata03/os/odata-json-format-v4.0-errata03-os-complete.html#_Toc453766627]]>
+    /// </remarks>
     public sealed class ODataResponseContent
     {
         /// <summary>
-        /// Gets or sets the URI to the metadata.
+        /// Gets or sets the context URI for the payload, unless 'odata.metadata=none' is specified in the request.
         /// </summary>
         [Newtonsoft.Json.JsonProperty("@odata.context", NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore, Order = 0)]
 #if NETSTANDARD2_0
@@ -27,16 +30,16 @@ namespace Net.Http.OData
         public string Context { get; set; }
 
         /// <summary>
-        /// Gets or sets the total result count.
+        /// Gets or sets the total count of members in the collection represented by the request when '$count=true' is specified.
         /// </summary>
         [Newtonsoft.Json.JsonProperty("@odata.count", NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore, Order = 1)]
 #if NETSTANDARD2_0
         [System.Text.Json.Serialization.JsonPropertyName("@odata.count")]
 #endif
-        public int? Count { get; set; }
+        public long? Count { get; set; }
 
         /// <summary>
-        /// Gets or sets the URI to the next results in a paged response.
+        /// Gets or sets the URL that allows retrieving the next subset of the requested collection, its presence indicates that a response is only a subset of the requested collection.
         /// </summary>
         [Newtonsoft.Json.JsonProperty("@odata.nextLink", NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore, Order = 2)]
 #if NETSTANDARD2_0
@@ -45,7 +48,7 @@ namespace Net.Http.OData
         public string NextLink { get; set; }
 
         /// <summary>
-        /// Gets or sets the value to be returned.
+        /// Gets or sets the payload containing the result of the OData request.
         /// </summary>
         [Newtonsoft.Json.JsonProperty(NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore, Order = 3)]
         public object Value { get; set; }
